backend/pyproject.toml
[project]
name = "intervue-backend"
version = "0.1.0"
description = "Backend for Intervue"
requires-python = ">=3.11"

dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]",
    "openai>=2.36.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.12",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "ruff>=0.15.12",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]",
]

[tool.uv]
package = false
